24.11.01 플레이어 공격 판정,몬스터 상태머신
Monster Script 상속으로 여러 몬스터
생성 가능
플레이어 무기를 판정으로 Ontrigger

+++ Assets/Scripts/Player/BasePlayerState.cs
@@ -9,14 +9,14 @@ public class BasePlayerState : ScriptableObject
     
     public int level;
 
-    public float hp;
-    public float maxHp;
+    public int hp;
+    public int maxHp;
 
-    public float stamina;
-    public float maxStamina;
+    public int stamina;
+    public int maxStamina;
 
-    public float attackPoint;
-    public float defencePoint;
+    public int attackPoint;
+    public int defencePoint;
 
     public ItemData currentWeapon;
     public ItemData currentArmor;
@@ -25,5 +25,5 @@ public class BasePlayerState : ScriptableObject
     public float currentExp;
     public float maxExp;
 
-    public float gold;
+    public int gold;
 }
